feat: CheckTeamAcitivity 구현

+++ app/src/main/java/com/example/kuroupproject/datas/TeamCheckData.kt
@@ -4,5 +4,5 @@ class TeamCheckData (
     var title: String,
     var place: String,
     var nowNumber: Int,
-    var totalNumber: Int,
+    var totalNumber: Long,
 )
